feat(web-server): external source + proxy validator

+++ backend/web-server/controllers/external-source.controller.ts
@@ -1,45 +1,53 @@
 import { KoaContext } from '../koa';
 import { ExternalSourceCommandDTO, ExternalSourceDTO } from '../../../shared/model/external-sources.model';
+import ValidatorInterface from '../../validators/validator.interface';
 
-const getExternalSources = async (ctx: KoaContext<void, Array<ExternalSourceDTO>>) => {
-  const externalSources = ctx.app.repositoryService.externalSourceRepository.getExternalSources();
-  ctx.ok(externalSources);
-};
+export default class ExternalSourceController {
+  constructor(private readonly validator: ValidatorInterface) {}
 
-const getExternalSource = async (ctx: KoaContext<void, ExternalSourceDTO>) => {
-  const externalSource = ctx.app.repositoryService.externalSourceRepository.getExternalSource(ctx.params.id);
-  ctx.ok(externalSource);
-};
-
-const createExternalSource = async (ctx: KoaContext<ExternalSourceCommandDTO, void>) => {
-  const command: ExternalSourceCommandDTO | undefined = ctx.request.body;
-  if (command) {
-    const externalSource = ctx.app.repositoryService.externalSourceRepository.createExternalSource(command);
-    ctx.created(externalSource);
-  } else {
-    ctx.badRequest();
+  async getExternalSources(ctx: KoaContext<void, Array<ExternalSourceDTO>>): Promise<void> {
+    const externalSources = ctx.app.repositoryService.externalSourceRepository.getExternalSources();
+    ctx.ok(externalSources);
   }
-};
 
-const updateExternalSource = async (ctx: KoaContext<ExternalSourceCommandDTO, void>) => {
-  const command: ExternalSourceCommandDTO | undefined = ctx.request.body;
-  if (command) {
-    ctx.app.repositoryService.externalSourceRepository.updateExternalSource(ctx.params.id, command);
-    ctx.noContent();
-  } else {
-    ctx.badRequest();
+  async getExternalSource(ctx: KoaContext<void, ExternalSourceDTO>): Promise<void> {
+    const externalSource = ctx.app.repositoryService.externalSourceRepository.getExternalSource(ctx.params.id);
+    if (externalSource) {
+      ctx.ok(externalSource);
+    } else {
+      ctx.notFound();
+    }
   }
-};
 
-const deleteExternalSource = async (ctx: KoaContext<void, void>) => {
-  ctx.app.repositoryService.externalSourceRepository.deleteExternalSource(ctx.params.id);
-  ctx.noContent();
-};
+  async createExternalSource(ctx: KoaContext<ExternalSourceCommandDTO, void>): Promise<void> {
+    try {
+      await this.validator.validate(ctx.request.body);
+      const externalSource = ctx.app.repositoryService.externalSourceRepository.createExternalSource(
+        ctx.request.body as ExternalSourceCommandDTO
+      );
+      ctx.created(externalSource);
+    } catch (error: any) {
+      ctx.badRequest(error.message);
+    }
+  }
 
-export default {
-  getExternalSources,
-  getExternalSource,
-  createExternalSource,
-  updateExternalSource,
-  deleteExternalSource
-};
+  async updateExternalSource(ctx: KoaContext<ExternalSourceCommandDTO, void>): Promise<void> {
+    try {
+      await this.validator.validate(ctx.request.body);
+      ctx.app.repositoryService.externalSourceRepository.updateExternalSource(ctx.params.id, ctx.request.body as ExternalSourceCommandDTO);
+      ctx.noContent();
+    } catch (error: any) {
+      ctx.badRequest(error.message);
+    }
+  }
+
+  async deleteExternalSource(ctx: KoaContext<void, void>): Promise<void> {
+    const externalSource = ctx.app.repositoryService.externalSourceRepository.getExternalSource(ctx.params.id);
+    if (externalSource) {
+      ctx.app.repositoryService.externalSourceRepository.deleteExternalSource(ctx.params.id);
+      ctx.noContent();
+    } else {
+      ctx.notFound();
+    }
+  }
+}
